// src/Models/State/state.ts
export class State{
    SrNo: number;
    Result: number;
    StateCode: string;
    StateName: string;
    CountryCode: string;
    CountryName: string;
    CreatedBy: string;
    CreatedOn: string;
    UpdatedBy: string;
    UpdatedOn: string;
}

export class Country{
    CountryCode: string;
    CountryName: string;
}